added comments; rearrangement; decent map validation and error check :white_check_mark:
