+++ init_map_utils_2.c
@@ -1,0 +1,114 @@
+#include "cub3d.h"
+
+/////////////////////////
+// utils for check_map //
+/////////////////////////
+// check the map for invalid characters
+void	check_chars(char **raw, t_input *data)
+{
+	size_t	j;
+	size_t	i;
+
+	j = 0;
+	while (raw[j])
+	{
+		i = 0;
+		while (raw[j][i])
+		{
+			if (!check_charset(raw[j][i], "10NEWS "))
+				error_exit(data, "Ivalid character on the map", 1);
+			++i;
+		}
+		++j;
+	}
+}
+
+// check that all rows are closed
+void	check_rows(char **map, t_input *data)
+{
+	size_t	j;
+	size_t	i;
+	int		closed;
+
+	j = 0;
+	closed = 1;
+	while (map[j])
+	{
+		i = 0;
+		while (map[j][i] && map[j][i] == ' ')
+			++i;
+		if (map[j][i] && map[j][i] != '1')
+			error_exit(data, "Unclosed map: rows", 1);
+		while (map[j][i])
+		{
+			if (map[j][i] == '1')
+				closed = 1;
+			else if (map[j][i] == '0')
+				closed = 0;
+			i++;
+		}
+		if (!closed)
+			error_exit(data, "Unclosed map: rows", 1);
+		++j;
+	}
+}
+
+// check whether a line was fully checked; if so, go to another one
+size_t	check_length(char **map, size_t i, t_input *data)
+{
+	size_t	j;
+	size_t	len;
+
+	j = data->j;
+	while (map[j])
+	{
+		len = ft_strlen(map[j]);
+		if (i < len)
+			break ;
+		++j;
+	}
+	data->j = j;
+	return (j);
+}
+
+// check that all columns are closed
+void	check_columns(char **map, t_input *data)
+{
+	size_t	j;
+	size_t	i;
+	int		closed;
+
+	j = 0;
+	data->j = 0;
+	closed = 1;
+	i = 0;
+	while (map[j])
+	{
+		while (map[j][i] && map[j][i] == ' ')
+			++j;
+		if (map[j][i] && map[j][i] != '1')
+			error_exit(data, "Unclosed map: columns", 1);
+		while (map[j] && map[j][i])
+		{
+			if (map[j][i] == '1')
+				closed = 1;
+			else if (map[j][i] == '0')
+				closed = 0;
+			j++;
+		}
+		if (!closed)
+			error_exit(data, "Unclosed map: columns", 1);
+		++i;
+		j = check_length(map, i, data);
+	}
+}
+
+void	check_map(t_map *map, t_input *data)
+{
+	check_chars(map->map, data);
+	printf("checking rows\n");
+	check_rows(map->map, data);
+	printf("checking columns\n");
+	check_columns(map->map, data);
+	printf("closed\n");
+}
